Refactor demo forms.
