+++ Accelist.FluentValidation.Blazor.sln
@@ -7,6 +7,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Accelist.FluentValidation.B
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Demo", "Demo\Demo.csproj", "{312A7B7D-69CE-4717-A3F4-36283B072147}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Demo.Forms", "Demo.Forms\Demo.Forms.csproj", "{C4A23AF3-D0F0-4700-A734-876B31141727}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -21,6 +23,10 @@ Global
 		{312A7B7D-69CE-4717-A3F4-36283B072147}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{312A7B7D-69CE-4717-A3F4-36283B072147}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{312A7B7D-69CE-4717-A3F4-36283B072147}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C4A23AF3-D0F0-4700-A734-876B31141727}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C4A23AF3-D0F0-4700-A734-876B31141727}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C4A23AF3-D0F0-4700-A734-876B31141727}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C4A23AF3-D0F0-4700-A734-876B31141727}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
