+++ Demo.Forms/Services/EmailCheckerService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 
-namespace Demo.Services
+namespace Demo.Forms.Services
 {
     public class EmailCheckerService
     {
