+++ Demo/Demo.csproj
@@ -13,6 +13,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\Demo.Forms\Demo.Forms.csproj" />
     <ProjectReference Include="..\FluentValidation.Blazor\Accelist.FluentValidation.Blazor.csproj" />
   </ItemGroup>
