+++ Demo/_Imports.razor
@@ -7,4 +7,5 @@
 @using Microsoft.JSInterop
 @using Demo
 @using Demo.Shared
-@using Demo.Models
+@using Demo.Forms.Models
+@using Demo.Forms.Services
